+++ src/core/sso/data.js
@@ -3,6 +3,6 @@ import Cache from '../../utils/cache';
 
 const cache = new Cache((...args) => webAPI.getSSOData(...args));
 
-export function fetchSSOData(id, cb) {
-  cache.get(id, cb);
+export function fetchSSOData(id, ...args) {
+  cache.get(id, ...args);
 }