+++ src/field/password/password_pane.jsx
@@ -17,12 +17,14 @@ export default class PasswordPane extends React.Component {
   };
 
   render() {
-    const { i18n, lock, placeholder, policy, strengthMessages } = this.props;
+    const { i18n, lock, placeholder, policy, strengthMessages, hidden } = this.props;
+    const hiddenCss = hidden ? ' auth0-lock-hidden' : '';
     return (
-      <div className="auth0-lock-input-block auth0-lock-input-show-password">
+      <div className={`auth0-lock-input-block auth0-lock-input-show-password${hiddenCss}`}>
         <PasswordInput
           value={c.getFieldValue(lock, 'password')}
           invalidHint={i18n.str('blankErrorHint')}
+          showPasswordStrengthMessage={!c.isFieldValid(lock, 'password')}
           isValid={!c.isFieldVisiblyInvalid(lock, 'password')}
           onChange={this.handleChange}
           placeholder={placeholder}
@@ -48,5 +50,6 @@ PasswordPane.propTypes = {
   onChange: PropTypes.func,
   placeholder: PropTypes.string.isRequired,
   policy: PropTypes.string,
-  strengthMessages: PropTypes.object
+  strengthMessages: PropTypes.object,
+  hidden: PropTypes.bool
 };